Refactor random generation to units package
