+++ pkg/utils/stringSliceUtils_test.go
@@ -1,4 +1,4 @@
-package k8s
+package utils
 
 import (
 	"testing"
