+++ pkg/utils/stringUtils_test.go
@@ -1,4 +1,4 @@
-package k8s
+package utils
 
 import (
 	"fmt"
